refactor: use grapheme clusters for glyph lookup
For latin script fonts, it works to assume a 1:1 mapping from each
char of a &str to a corresponding glyph. But, that assumption fails
badly for emoji and other scripts which can use more than one
codepoint to represent one glyph.

This commit changes the fonts to look up glyphs from grapheme
clusters using binary search with a precomputed index table of
murmur3 hashes.